build: disable zlib while making version files
cmake is invoked to generate version, AST, ASDL, etc. files
in-source that are in turn used to generate the tarballs.
The MacOS GitHub Actions CI for some reason is unable to
detect zlib when called from
build_tmp.xsh>create_source_tarball.xsh>cmake ..
so we disable zlib for the temp file creation stage.

NOTE: unable to replicate on local MacOS Monteray 12.5