CI: test Minpack's primes example
We reworked this example to compile with LFortran. This example was
written by me, so it is up to us how we do it, and in this test we are
focusing on Minpack, not other unrelated features.

This example uncovered issue #1224, so for now we modified Minpack to
use x**0.5 instead of sqrt(), in order to get exactly the same results
with LFortran and GFortran. After we implement sqrt() directly (not
using x**0.5), we can revert the Minpack workaround.