[parser] enh: allow whitespace in Format specification
Blank characters may precede the initial left parenthesis of the
format specification and additional blank characters may appear
at any point within the format specification, with no effect on
the interpretation of the format specification.

e.g.; `(I 1 2)` is same as `(I12)` for format specification
and are both valid.

Earlier, we used to raise tokenizer error because of no support
for it

issue was initially raised in #4040