[llvm] Fix passing character to subroutine with bindC convention
Under bindC conventions, character types are passed as byte stream.
Under normal conventions, they are represented as a pointer to a byte stream.

As a consequence, when calling a bindC function, they need to be
referenced through a load when packing the arguments, and their address
needs to be taken when unpacking them.