CI: do not run simplifier for llvm_wasm tests
Given:

* The simplifier is tested with LLVM in other tests
* We only changed the backend in LLVM to use WASM

And assuming:

* No bugs in LLVM

We do not need to test simplifier with llvm_wasm, because it should not
trigger any bugs. If it does, those would be LLVM bugs, or integration
bugs, but those are rare. For now the main thing is to ensure the
simplifier pass works and it works with all the other passes and with
LLVM. We already test that. We do not need to test every combination of
the simplifier and other options.

Once simplifier becomes ready, then we can start testing it with all
combinations. Until then we just need to ensure it works with one set of
options, not all.