Mimic old common-block-object parsing behavior
Before the introduction of the direct parsing of a _common-block-object_,
the object name and dimensions were stored in the `r->m_initializer`
expr, and decoded in ast_common_visitor.  This change mimics that
behavior by converting the `var_sym_t` into the form of the expr that
would have been created with the old parser.  This allows us to test
parser changes without substantive changes to AST->ASR code.