Switch to new common-block encoding
The AST now encodes common blocks as a block-name `var_sym_t`, followed
by `var_sym_t` for each common-block-object specification.  This allows
the blank common block to appear in the middle of a `COMMON` block list,
using the `//` notation.

This commit also changes the blank block name to something (`blank#block`)
that is not a valid Fortran identifier, to prevent clashes with user
specified block names.