fix: set `ArrayPhysicalCast` type correctly when passing an `allocatable` array as argument to a fixed-size array
Looking at the `asr_verify` code for `Allocatable_t` with the `simplifier`
pass, we see that it requires an `allocatable` type to have a deferred length
to work. This is the required behavior also, but for `ArrayPhysicalCast`, we
were setting the `ttype` wrapped inside `Allocatable_t` causing issues. As it
turns out, the implementation before the `simplifier` pass had handled this with
`Allocatable_t`, but `simplifier` pass is stricter in this regard.