Make characters comparisons unsigned.
I discovered that lfortran was doing signed character comparisons on
many backends as part of #6188. I'm updating ASR, LLVM and C backends to
use unsigned string comparisons.

For the LLVM backend, the various intrinsics for string comparisons can
be updated to compare characters as unsigned values. Similarly, the C
backend can be updated to invoke a new intrinsic which mirrors `strcmp`
but which does unsigned comparisons.

`libasr` also requires a string comparison which respects Fortran
semantics. I'm adding a new `str_compare_fortran` helper to
`string_utils` which can be used by ASR passes.

The C++ backend is not updated. However, it seems to have significant
issues with operator precedence, and I can't get the emitted C++
programs to compile without large changes. C++ should also be able to
call the compiler intrinsics rather than depending on the comparison
operators for `std::string`.

The Julia backend does not appear to be affected, since Julia already
compares characters as unsigned values.

The WASM backend does not appear to be affected, `emit_string_cmp` is
already loading the characters within the string as unsigned integers.