feat: implement type to string functionality for `ASR::StructType_t`
We create a set of all struct names and a map from `ASR::ttype_t*` to `std::string*`. The idea behind this implementation is that a `ASR::ttype_t*` pointer must point to the same type declared on heap, and thus get the appropriate name from the map when requested for.

Signed-off-by: Saurabh Kumar <developer.saurabh@outlook.com>