Fix LLVM codegen for DATA on equivalenced arrays
For pointer types (equivalenced arrays), don't set m_value/m_symbolic_value
on the Variable as it causes null pointer issues in LLVM codegen. Instead,
generate an assignment statement to initialize the array at runtime.

This is the second part of the EQUIVALENCE + DATA fix - the first commit
fixed the down_cast assert, this fixes the subsequent LLVM error.

Refs #9115, refs #5529